+++ docs/changelog.md
@@ -13,6 +13,7 @@ category: Administration
 
 * docs: add frequently asked questions to online documentation
 * feat: support estimate fields that are select lists or radio buttons (#120)
+* refactor: improve code to persist estimate depending on field type 
 
 ### [20.12.1] - 2020-12-07