+++ docs/changelog.md
@@ -16,6 +16,7 @@ category: Administration
 * test: add tests for getting started page (#123)
 * feat: add fallback colors for getting started page (#123)
 * feat: include references to documentation and service desk (#123)
+* docs: add documentation url to Maven POM
 
 ### [21.01.0] - 2021-01-08